+++ resources/views/backend/product/index.blade.php
@@ -6,8 +6,8 @@
             Sản phẩm
         </h1>
         <ol class="breadcrumb">
-            <li><a href="#"><i class="fa fa-dashboard"></i> Trang chủ</a></li>
-            <li class="active">Sản phẩm</li>
+            <li><a href="{{asset('/admin')}}"><i class="fa fa-dashboard"></i> Trang Chủ</a></li>
+            <li class="active">Danh sách Sản phẩm</li>
         </ol>
     </section>
 
@@ -17,7 +17,15 @@
             <div class="col-xs-12 table-responsive">
                 <div class="box">
                     <div class="box-header with-border">
-                        <h3 class="box-title">Danh sách Sản phẩm</h3>
+                        @if(\Auth::user()->role_id == 1) <!-- kiểm tra tài khoản có phải là admin ko, nếu là admin thì show combobox filter -->
+                        <div class="form-group" style="width: 150px;float: left;margin: 0">
+                            <select class="form-control" id="filter_type" name="filter_type">
+                                <option {{ $filter_type == 1 ? 'selected' : '' }} value="1">Tất cả</option>
+                                <option {{ $filter_type == 2 ? 'selected' : '' }} value="2">Đang Sử Dụng</option>
+                                <option {{ $filter_type == 3 ? 'selected' : '' }} value="3">Thùng rác</option>
+                            </select>
+                        </div>
+                        @endif
                         <a href="{{ route('admin.product.create') }}" class="btn btn-primary pull-right"><i class="fa fa-plus" aria-hidden="true"></i></a>
                     </div>
                     <!-- /.box-header -->
@@ -41,7 +49,7 @@
                                         @if($item->image && file_exists(public_path($item->image)))
                                             <img src="{{ asset($item->image) }}" width="100" height="75" alt="">
                                         @else
-                                            <img src="{{ asset('upload/404.png') }}" width="100" height="75" alt="">
+                                            <img src="{{ asset('frontend\Img404.png') }}" width="100" height="75" alt="">
                                         @endif
                                     </td>
                                     <td>{{ $item->name }}</td>
@@ -50,9 +58,15 @@
                                     <td>
                                         {!! $item->is_active == 1 ? '<span class="badge bg-green">ON</span>' : '<span class="badge bg-danger">OFF</span>' !!}
                                     </td>
-                                    <td>
+                                    <td class="action" >
                                         <a href="{{ route('admin.product.edit', ['product' => $item->id]) }}"><span title="Chỉnh sửa" type="button" class="btn btn-flat btn-primary"><i class="fa fa-edit"></i></span></a>
-                                        <span data-id="{{ $item->id }}" title="Xóa" class="btn btn-flat btn-danger deleteItem"><i class="fa fa-trash"></i></span>
+                                        @if($item->deleted_at == null)
+                                            <span  data-id="{{ $item->id }}" title="Xóa" class="btn btn-flat btn-danger deleteItem"><i class="fa fa-trash"></i></span>
+                                            <span style="display:none;" data-id="{{ $item->id }}" title="Khôi phục" class="btn btn-flat btn-warning restoreItem"><i class="fa fa-refresh"></i></span>
+                                        @else
+                                            <span style="display:none;" data-id="{{ $item->id }}" title="Xóa" class="btn btn-flat btn-danger deleteItem"><i class="fa fa-trash"></i></span>
+                                            <span  data-id="{{ $item->id }}" title="Khôi phục" class="btn btn-flat btn-warning restoreItem"><i class="fa fa-refresh"></i></span>
+                                        @endif
                                     </td>
                                 </tr>
                             @endforeach
@@ -102,6 +116,53 @@
                         });
                     }
                 });
+            });
+
+            // bắt sự kiện nút khôi phục
+            $('.restoreItem').click(function () {
+                var id = $(this).attr('data-id');
+                Swal.fire({
+                    title: 'Bạn có muốn khôi phục ?',
+                    text: "Dữ liệu sẽ được khôi phục lại danh sách Sản Phẩm",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'OK'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        $.ajax({
+                            url : '/admin/product/restore/'+id,
+                            type: 'POST',
+                            data: {},
+                            success: function (res) {
+                                if(res.status) {
+                                    Swal.fire(
+                                        'Thông báo !',
+                                        'Khôi phục thành công',
+                                        'success'
+                                    )
+
+                                } else {
+                                    Swal.fire(
+                                        'Thông báo !',
+                                        'Có lỗi xảy ra',
+                                        'error'
+                                    )
+                                }
+                            },
+                            error: function (res) {
+                            }
+                        });
+                    }
+                });
+            });
+
+            // bắt sự kiện combobox filter
+            $('#filter_type').change(function () {
+                var filter_type = $('#filter_type').val();
+                // sử dụng window.location.href để truyền dữ liệu filter và chuyển trang
+                window.location.href = "{{ route('admin.product.index') }}?filter_type="+filter_type;
             });
         });
     </script>
